Added new "public String[] getValues()" method to ListValueSource interface and all implementations.
Added a new ListValueSource type called "query-cols:abc" (QueryColumnsListValue) that will execute a query and return a strings list of each column (only the first row). This value source is useful in <field.grid captions="query-cols:xyz"/> or anywhere else a list needs to be inferred from a series of columns instead of rows.