Changed the caption attribute of all dialog fields from a static string to SingleValueSource. This will allow all captions to be dynamic (when needed).
Added a "captions" attribute (a ListValueSource) to the GridField that would override the captions specified in the grid row. Captions is a list value source so one could use "strings:a;b;c" to create a static list or use "query:abc" that would run a named query and set the captions to the first column of each row returned; or, "query-cols:xyz" that would run a named query and set the captions to be the values of each column in the first row returned from the query.