Created a new "mode" for dialogs called auto-execute or immediate-execute mode. Basically, if a dialog is called and a URL parameter called "_d_exec" is set to "1" or "yes" then the dialog never actually goes into input mode unless there is a validation error.
The only requirement is that when _d_exec is called, all parameters that would have been entered by the user be supplied as part of the URL query string. For example, if there are fields in the form called "field_00" and "field_01" then

http://<app>/mydialog.jsp?_d_exec=1&_dc.field_00=value&_dc.field_01=value2

would call the page with the dialog and set the form field called "field_00" to "value" and "field_01" to "value2" and automatically try to execute the dialog without requiring the user to enter data. If the fields don't have any validation errors then execution of the dialog starts immediately.

If there are grids or composite fields, then care must be taken to ensure that the entire qualified field name is used "_dc.<grid-name>_<row-name>_<field-name", etc. To be sure that you have the right names, run the dialog in regular mode, and look at the HTML source to get fully qualified field names with the starting _dc.