[Version 1.2.0 Build 62]
Added new feature to allow QuerySelectDialog tags (<select-dialog> in <query-definition>) to specify dynamic sort elements. Previous to Build 1.2.0.62, the <order-by> tag in the <select> tag of the <select-dialog> tag took only a "field" attribute that was a static string. Now, it takes a single-value-source (so it can be static or dynamic).

The following example demonstrates the use of a <field.select> that allows a user to select a fixed-sort type that allows the instruction <order-by field="form:order_by"/> to dynamically pick the sort field.

NOTE: if multiple sort types are needed at this time, they must be specified as separate <field.select> elements with corresponding <order-by> elements.

<select-dialog name="basic" heading="Basic" select="default">
	<field.text query-field="grantee_name"/>
	<field.select name="order_by" caption="Order by" choices="Vendor=vendor_code;Grantee Name=grantee_name"/>
	<select>
		<display field="vendor_code"/>
		<display field="grantee_name"/>
		<order-by field="form:order_by"/>
		<condition field="grantee_name" comparison="equals" value="form:grantee_name"/>
	</select>
</select-dialog>