[Version 1.2.1 Build 0]  -- StatementManager re-factorization
1) Added ability to create pseudo-dynamic SQL statements by allowing value source replacements inside SQL statements. This will allow basic replacements of ${xxx} where xxx is a single-value or list-value source. This change was needed to support the SQL "in" clause. For example,

<statement name="list">
	select *
	from table
	where column0 = ? or column1 in (${param-list:1})
	<params>
		<param value="request:state"/>
		<param values="form:abc"/>
	</params>
</statement>

In the above scenario, the SQL statement would be dynamically modified to read the values from a _list_ context (notice param "values" attribute instead of "value" attribute). The only time this type of variable replacement would be needed for sure is to support the "in" clause -- but, other ideas for use are welcome.

2) Moved the StatementInfo class out of StatementManager (it was getting too big). Created a new StatementParameter class to manage statements paramters.