[Version 1.2.1 Build 0]
Added experimental code to xaf:dml task and tag that would allow "insert or update" actions. Basically, two new attributes were added to provide "check" values that if they return true, would cause the "insert" or "update" happen. NOTE: the single value source used in either insertchk or updatechk MUST return an integer value (0 or 1).

For example,

<xaf:dml
  command="insert-or-update"
  table="helpme"
  columns="a=string:a,b=string:b,c=string:c"
  debug="yes"
  insertchk="request-attr:INSERTING"
  where="abc"/>

The above line would do an INSERT if the value of the requestAttribute named INSERTING was non-null and non-zero. If the value is zero or null, then an UPDATE would take place.

The opposite is true if the following were used (note here we are using the NEW single-value-source method of calling "query:"). In this case, XAF will run the query called "column-exists" and if it returns a single non-null non-zero value it will be updated.

<xaf:dml
  command="insert-or-update"
  table="helpme"
  columns="a=string:a,b=string:b,c=string:c"
  debug="yes"
  updatechk="query:column-exists"
  where="abc"/>

SPECIAL NOTE: this change requires, along with the update, the following lines added to the xaf.tld file in WEB-INF/tld <xaf:dml> tag:

<attribute>
	<name>insertchk</name>
	<rtexprvalue>false</rtexprvalue>
</attribute>
<attribute>
	<name>updatechk</name>
	<rtexprvalue>false</rtexprvalue>
</attribute>