[Version 1.2.1 Build 12] Added initial support for defining dialog skins in XML. For example, now in the dialogs.xml one can specify most easily-modifiable style parameters in the source:
NOTE: Setting the "default" skin changes the UI for all dialogs. To see all of the attributes, please refer to com.xaf.skin.StandardDialogSkin at this time. Will be updating the Tagdoc soon.

<xaf>
	<dialog-skin name="default">
		<outer-table-attrs>
			cellspacing='1' cellpadding='0' bgcolor='black'
		</outer-table-attrs>
		<frame-head-row-attrs>
			bgcolor='blue'
		</frame-head-row-attrs>
		<include-pre-script src="/test/script.js"/>
		<include-post-script src="/test/script-post.js"/>
		<prepend-pre-script>
		value = "this";
		</prepend-pre-script>
		<append-post-script>
		alert(activeDialog);
		</append-post-script>
	</dialog-skin>

	<dialogs package="app">
		<dialog name="login" heading="Please Login">
.....


NOTE:

there is also an attribute called "class" that can be used in <dialog-skin> that would use the provided classname as the base class. This allows a combination of Java and XML to be used for manipulating skins.

Support for JavaScript calls and external CSS and include files has also been added.