[Version 1.2.5 Build 51]
Added feature to allow multiple order-bys to be provided as a value source.
Reported by Moorthy.

In the following example, the "order_by" field has the choices Vendor,
Grantee Name, and Grantee, Vendor. Before Build 52, when there were multiple
items in a value-source for an order-by field the XAF generated an error.
Now, it will accept any number of columns for a dynamic order-by clause.

<select-dialog name="basic" heading="Basic" select="default">
	<field.text query-field="grantee_name"/>
	<field.select name="order_by" caption="Order by" choices="Vendor=vendor_code;Grantee Name=grantee_name;Grantee, Vendor=grantee_name,vendor_code"/>
	<select>
		<display field="vendor_code"/>
		<display field="grantee_name"/>
		<order-by field="form:order_by"/>
		<condition field="grantee_name" comparison="equals" value="form:grantee_name"/>
	</select>
</select-dialog>