[Version 1.2.5 Build 51]
Added new Data Command infrastructure which allows the same dialog code to be used in different roles. For instance, the same dialog can act as a "add",  "edit", "delete", or "confirm" dialog. When adding, all fields are visible and editable. In update mode, the primary key is read-only and all other fields are editable. In delete and confirm mode, all fields are read-only. Requested by Jeevan and Moorthy.

When calling a dialog, set
   /abc/def.jsp?data_cmd=mode

where "mode" can be "add", "edit", "delete", or "confirm". NOTE: special thought needs to be given to how the confirm and delete messages/headings should appear (for UI consistency). Talk to me if you have any ideas on this matter.