[Version 1.2.7 Build 71]
Changed methods in the DatabaseContext interface so that each method now accepts a ValueContext parameter. Using this parameter, and new methods like translateDataSourceId(), NBCS and other enterprise-scale applications can override dataSourceIds so that they can "sniff" the environment and auto-select a data source from a bank of IDs. This change was needed to support multiple regional databases like CDIS/DMIS.