[Version 1.2.7 Build 73]
Added capability to use multiple skins and specify output destiations for <xaf:query> and <exec-statement> tags. In the example dialog shown below, there is a single SQL statement that is executed but the dialog is run to find out HOW and WHERE the user would like to see their results.

The important change to note are the two attributes in <exec-statement> (same thing works in <xaf:query> in JSP):

<exec-statement skin="form:skin" destination="form:destination">
<xaf:query skin="form:skin" destination="form:destination">

The "skin" and "destination" are single value sources which means their values can come from almost anywhere (in the example we use "form"). Skin may be any report skin enumerated in com.xaf.skin.SkinFactory.java and destination may be "browser", "file", or "email". Realize that e-mail doesn't work yet but it should in the next few days or week.


<dialog name="testdest" heading="Test CSV" loop="no">
	<field.select name="skin" caption="Skin" required="yes" choices="Report=report;CVS=text-csv;Tab=text-tab"/>
	<field.select name="destination" caption="Destination" required="yes" choices="Browser=browser;File=file;E-mail=email"/>
	<execute-tasks>
		<exec-statement skin="form:skin" destination="form:destination">
			select statename + ' (' + statecode + ')', statecode as "Value"
			from state
			order by statename
		</exec-statement>
		<!-- <exec-redirect url="config:something"/> -->
	</execute-tasks>
</dialog>