[Version 1.2.8 Build 18] Added new "application execution environment" variable in web.xml that can be set to one of the following variables: "Development", "Testing", or "Production". When the value is "Production" or "Testing" some special things will start to happen. First, the automatic XML reloading will be turned off so that in production the XML files are not checked for changes on each refresh of a page. Another change which is forthcoming will be to limit the amount of SQL execution statistics that is captured in-memory when the execution environment is "Production". The on-disk captures will of course not be limited in any environment.
To specify the execution environment, add the following to WEB.XML:

	<!-- setup execution environment -->
	<context-param>
	  <param-name>app-exec-environment</param-name>
	  <param-value>Production</param-value>
	</context-param>

Again, the values of XXXX in <param-value>XXXX</param-value> above can be
* Production
* Testing
* Development (this is the "default" if value is NOT set at all (null))

In Build 18 now, in various places in ACE you will now see and options block
(like in Dialogs, SQL, etc) that shows whether automatic XML reloads are turned
on or off.

Also, on the ACE home page right above the version number you will see the
"XXX Environment" string where XXX will be either Production, Testing, or
Development.

If developers would like to use the environment string in their own
applications, they can do so by using new methods in the following class:
com.xaf.config.ConfigurationManagerFactory. The methods are self-explanatory.

In this Build, it looks like there were many changes made; however, most of the
changes are very simple (one-liners) to check for the environment string before
doing something. The majority of the code changes were also in ACE to show the
value of the new environment string.