[Version 1.2.8 Build 26] ** IMPORTANT ** This commit is a partial commit in terms of functionality. Unless you're ready to try some new code, don't bother updating until Build 27 or 28.
This Build (26) is the first of a few builds that will make it easier to populate dialog data and reduce the amount of work necessary to create add/update/delete data dialogs. This build allows for the DialogContext to be subclassed (it used to be a final class before). Subclasses of DialogContext ("data beans") can be automatically generated by using the new "dc-class='generate'" attribute of the <dialog> class. More information and usage examples will be provided in the next few days.