[Version 1.2.8 Build 32] Added ability to easily attach subclassed DialogDirector to individual dialogs. There is a new attribute called "director-class" which works just like "class" and "dc-class" for dialogs:
<dialog name="abc" director-class="my.subclassed.Director">

If a class by the name my.subclassed.Director (or any other name) is found (and it must extend DialogDirector) then that class is used instead of com.xaf.form.DialogDirector. ACE will also show the subclassed Director in the dialogs list (or an exception if the class is not found). The subclassed director has full access to the XML attributes and children elements of the <director> element. For example, programmers could override the following:

package my.subclassed;
public class Director extends com.xaf.form.DialogDirector
{
    // stuff...

    public void importFromXml(Element elem)
    {
        super.importFromXml();
        // grab your own XML definitions since "elem" is the <director> element      }

    public String getControlHtml(DialogContext dc)
    {
        String superHtml = super.getControlHtml(dc);
        // render your own HTML in any way you like
        return superHtml + " my special HTML";
    }
}