[Version 1.2.8 Build 32] Added ability to easily register subclassed ColumnDataCalculator classes from XML. The tag is called <register-column-data-calc> and is placed into the sql/statements.sql file at the level right underneath <xaf> (same level as <sql-statements> and <<query-defn>):
<xaf>
   <register-column-data-calc name="my-sum" class="my.calc.sum"/>
   ...
   <sql-statements package="blah.blah"> ...
   <query-defn name="blah.blah>...
</xaf>

If the registration is successful, then the new calculator will appear in ACE -> Application -> Factories -> Report Components page. If unsuccessfull, then there will be an error message that appears at the bottom of the ACE -> Database -> SQL page.

If the registration was successful, then the new class can be used just like "sum" or "count" in a report:

<report>
  <column calc="my-calc">