[Version 1.2.8 Build 35] Added rudimentary support for file upload fields. This is accomplished by adding a new <field.file> tag created by Joe Russ (with minor modifications by myself). The <field.file> is derived from TextField but really only supports the "caption", "name", and "size" attributes (due to browser limitations). Please read http://www.cs.tut.fi/~jkorpela/forms/file.html for a very thorough introduction to uploading files using browsers.
When a field of FileField type (<field.file> tag) is added to a dialog, an automatic flag is set in the dialog that the StandardDialogSkin now checks to appropriately set enctype="multipart/form-data" in the <form> tag that is generated by the skin.

However, given the multipart/form-data is not supported by the default servlet Request object, the programmer needs to add appropriate support for parsing the multipart form data. I would recommend using the utilities provided by the Oreilly Servlet library found at http://www.servlets.com/cos/index.html (use the MultipartParser class to prevent having to write your own). Soon this parser library will be integrated with the XAF.