[Version 1.2.8 Build 38]
Added new DatabasePolicy interface that sets up a generic mechanism to control DBMS-specific behavior. Unlike the DatabaseContext interface, which manages Connection-specific and Driver-specific behavior, the DatabasePolicy will provide programmers with a generic interface to help shield database-specific SQL.

The only three methods in this new DatabasePolicy interface at this time help with generating auto-inc and sequence values for SQL insert statements. Please see the javadoc for documentation on the three methods.

The OracleDatabasePolicy is the only database who's policy has been created, please see the .java file for more information.

To help with debugging, the database product, database version, and specific class instantiated for a particular DatabasePolicy for each JDBC driver is now shown in the ACE Data Sources page.