[Version 1.2.8 Build 38]
Added new DialogDataCmdExpr single value source that is identified by "create-data-cmd-heading". This is a very useful SVS that be used in <dialog> heading tags for automatically creating a heading based on a data command. For example,

<dialog heading="create-data-cmd-heading:Account">

will automatically set the heading to "Add Account" when the data command is "add", to "Edit Account" when the data command is edit, and "Delete Account" or "Confirm Account" if the data command is delete or confirm. At this time, the SVS behavior pretty simple: all it does is to take a title-cased command name and prepends it to the string provided to the SVS.