[Version 1.2.8 Build 38] Replaced the class AbstractTask with the class BasicTask. The primary new feature is the addition of child-tasks and the ability for tasks to execute conditionally based on the current dialod data command.
EXAMPLE:

<populate-tasks data-cmd="edit,delete">
   ... whatever ...
</populate-tasks>

The above block will only execute if the dialog data command is edit OR delete.

<execute-tasks data-cmd="add">
   ... whatever ...
</execute-tasks>

The above block will only execute if the dialog data command is add.

<execute-tasks data-cmd="add,edit">
   ... whatever ...
   <exec-task ...>
   <exec-task data-cmd="edit">
   <exec-task ...>
</execute-tasks>

The above block will only execute if the dialog data command is add OR edit, but the second task will only execute when the data command is edit.

<execute-tasks data-cmd="delete">
   ... whatever ...
</execute-tasks>

The above block will only execute if the dialog data command is delete.