[Version 1.2.8 Build 39] Changed the method signatures for the com.xaf.report.ColumnDataCalculator interface. Now, each addValue method accepts the following parameters:
    public void addValue(ReportContext rc, ReportColumn columnInfo, long rowNum, Object[] rowData, double value);

    public void addValue(ReportContext rc, ReportColumn columnInfo, long rowNum, Object[] rowData, String value);

The reason for this addtion was to allow the column data calculators to look at their environment (using ReportContext), it's column definition (using ReportColumn), the current rowNum, or any of the other columns in the row (using rowData). The actual data that should be used for computation is contained in the value parameter.

** Another addition was the ability to specifiy a class name for the calc="xxx" attribute. For example:

<report>
   <column heading="ID" calc="report.calc.MyCalc"/>

Now, calc can be a built-in calculator like "sum" or "count" or it can be one that was registered using <register-column-data-calc> or it can be an actual class name that will be searched for in the classpath using Class.forName().

For one-offs or special cases use calc="pkg.class.name" format. For standard ones or those that will be used over and over, use the <register-column-data-calc> specification.