[Version 1.2.8 Build 39] Added ability for the "query:" single value source to take dynamic parameters. Now, instead of just being able to say
<field.select choices="query:person.active-org-memberships"/>

you can now say

<field.select choices="query:person.active-org-memberships?session:active-user"/>

Using the second method, you can pass in one or more (comma-separated) single value sources that can be used to override the parameters to the named query. In the above example, the statement called "person.active-org-memberships" will be looked-up and called using the parameter "session:active-user" and the results will be placed in the choices list. This new method really works best for string parameters or with databases (like ORACLE) that can convert strings to numbers automatically.