[Version 1.2.8 Build 41] Added two new convenience methods:
boolean stmtRecordExists()
static boolean sqlRecordExists()

The objective of these methods is to simply return true if any row is found in a query's ResultSet. The idea is to take either a static (XML) sql statement or a dynamic sql statement, execute it and see if there are any rows in the result set.