[Version 1.2.8 Build 49] Added capability to define and use field and dialog templates in dialogs.xml (DialogManager). What this feature effectively does is allows common fields and dialogs to be created and stored in a "templates pool" and then re-used. For example, if you have a field type that is commonly used over and over with the same validation or caption or feature, you can create a new field type subclass or use the new templating feature. Basically, templating allows for re-use of fields without resorting to Java.
DEFINING THE TEMPLATES

<xaf>
<templates>
  <template id="edit-delete-conditional">
    <conditional action="apply-flag" flag="invisible" data-cmd="edit,delete"/>
  </template>

  <field.text id="field.string" name="string" caption="String" reg-ex="this | that" template="edit-delete-conditional"/>

  <dialog id="dialog.common" retain-params="org_id" class="app.dialog.org.Registration">
    <field.text name="org_code" caption="Account Code" required="yes"/>
    <field.text name="org_name" caption="Name" required="yes"/>
    <field.text name="org_abbrev" caption="Abbreviation"/>
  </dialog>
</templates>

Note that templates are defined in a new top-level tag called <templates>. It can optionally have a package attribute like <templates package="abc">. If you specify a package then all template IDs are prefixed with that package and looked up with the package prefix.

Each element underneath the <templates> element is considered a "template" and is stored by the name provided in the "id" attribute. Any element may be stored as a template but the way a template is referenced determines whether the entire template or just the template's children are copied when used.

USING THE TEMPLATES

<xaf>
<templates ...>
<dialogs package="org">
  <dialog name="templatetest" template="dialog.common">
    <field.string/>
    <field.text template="field.string"/>
    <include-template id="dialog.common"/>
  </dialog>

  <dialog name="reg" heading="create-data-cmd-heading:Account" retain-params="org_id" class="app.dialog.org.Registration">
    <field.debug visible="no"/>
    <field.string/>
    <field.text name="org_code" caption="Account Code" required="yes"/>
    <field.text name="org_name" caption="Name" required="yes"/>
    <field.text name="org_abbrev" caption="Abbreviation"/>
    .......

Each <dialog> and <field.xxx> may now have an attribute called "template" that will lookup the template and "inherit" the items that are in the template. What this means is that all of the attributes that are not overriden by the template caller are copied and all of the children of the template are copied into the caller. More than one template may be copied by stating template="A,B,C,etc".

Another way of getting items from a template is to use the <include-template> element which acts like an <include file> in the sense that it does not copy attributes or main elements of a template, just children of the template into the caller. This is useful to create "snippets" of reusable dialogs and fields.

Yet another method is to create a template named "ABC" and then reference it by creating an element called <ABC> inside a dialog (as a field). For example, the <field.text id="field.string" ...> template declaration in the above example means that we want to be able to reference the template by declaring <field.string name="abc" ...> instead of <field.text template="field.string">. This is provided for convenience and is the more readable form for overriding fields.

NOTE: (1) templates can inherit other templates as well and (2) you can see the templates that are being inherited/copied/extended by review the Dialog details in ACE.