[Version 1.2.8 Build 52] Added features so that the allow-null="no" attribute will properly work in nested <condition> elements in <select> tag of <query-defn>.
In Build 50, the nested <condition> was added but it didn't automatically remove it's child conditions that were dynamic (should be removed if null). Now, the behavior is more consistent: all child conditions that should be remove are properly removed. Moreover, now if all children of a nested condition are removed, the entire nested condition is also removed.

This new feature is considered experimental and only basic testing has been performed at this time. Everything should work but the files are being committed so that users can perform further testing and provide suggestions for improvements.