[Version 1.2.8 Build 56] Introduced support specifying column values as child elements for the "enumeration" tables. For example, the following: <enum id="1" btypeid="O" type="O" active="Y"   java-constant="RBC_OPTIMAL_OPOS" name="RBC"/>
can be expressed as:
<enum id="1" name="RBC" java-constant="RBC_OPTIMAL_OPOS">
            <btypeid>O</btypeid>
            <type>O</type>
            <active>Y</active>
</enum>

IMPORTANT: column attributes can be expressed as child elements of the
'enum' element (meaning non-column attribute such as 'java-constant'
cannot be expressed as a child element).  Also, when child elements of
'enum' are defined, there can be no text value for the enum element.
For example, the following which assigns the text 'New' to the default
'caption:' column:

<enum id="1">New</enum>

cannot be expressed as:
<enum>
    <id>1</id>
    New
</enum>

If there is a child element(such as 'id' as shown above), then the text value
between the 'enum' start and end tag must also be defined as either
an attribute or a child element:

<enum>
    <id>1</id>
    <caption>New</caption>
</enum>