[Version 1.2.8 Build 60] Added new popup calendar functionality. Now, the <field.date> and <field.duration> tags have a new optional attribute called "popup-calendar" which can be set to "yes" to allow a popup calendar.
For example:

<field.duration ... popup-calendar="yes"/>
<field.date ... popup-calendar="yes"/>

When set to yes, a configurable icon appears next to the field which, when clicked, launches a popup window with a small calendar displayed in the window. The popup window allows selection of a date which, when clicked, fills in the date field control.

The script that controls the calendar and the icon are both configurable by adding the following properties to configuration.xml:

-- framework.shared.dialog.field.calendar.js-src
(default is "/shared/resources/scripts/calendar.js")

-- framework.shared.dialog.field.calendar.image-url
(default is "/shared/resources/images/navigate/calendar.gif")