[Version 2.0.3 Build 2] Changed the com.netspective.sparx.util.value.ListSource.java class so that it now implements the SingleValueSource interface as well as the ListValueSource interface. This minor code change is actually a big logical change and will now allow all List value sources to double as single value sources. This change was necessitated by the bug that Joe Russ found in the following case:
-- in a query definition <select-dialog> <select> tag:

<condition field="record_status_id" allow-null="no" comparison="in"
value="101,102" connector="and"/>

in the above case, the in comparison was looking for a ListValueSource but value="101,102" will always be a single value source. The simple fix should have been:

<condition field="record_status_id" allow-null="no" comparison="in"
value="strings:101;102" connector="and"/>

However, when the above was used, the value attribute got a class cast exception because it was expecting a single value source (not a list value source).

This change in 2.0.3 Build 3 changes all list value sources so that they also implement the single value source interface. Most of the unit test ran without problems but I am providing this fix quickly so that it can make it into the IPMS and DMI integration test phases.

If anyone finds anomalies with respect to list value sources acting oddly, please bring it to my attention immediately.