[Version 2.0.4 Build 1] Added new timeout method and changed rules for when connections are freed in query definitions. These are for the special situations in IPMS.
The rules are now:
1) If a result set is scrollable and there is only one page of query data then the scroll state is immediately removed and the connection is closed at the end of the page.
2) If a result set is scrollable and all data is requested to fit in one page by the programmer, then the scroll state is immediately removed and the connection is closed at the end of the page.

Also, In ResultSetScrollState (and by inheritance in QuerySelectScrollState) there are now member variables that track usage of the scroll state and a new method called isActive(int timeout). The idea is that now in PageTag or some other regular procedure, end-user applications can check the active scroll state for a user and run something like state.isActive(1000) to see if the scroll state has been used within the past 1000 milliseconds. Presumably, if the state is inactive for a while, it should be closed.