[Version 2.0.4 Build 3] Added new HypersonicSQL/HSQLDB (http://hsqldb.sourceforge.net) embedded Java database as a supported DatabasePolicy and DDL generator. The sample CURA application will be updated to use this database as the default database in order to make packaging the sample application easier. Thanks Joe!
The templates/xif/datatypes.xml file is the datatypes files which has the few things we found that HSqlDB needed different from ANSI, ORACLE, or SQL Server (should be merged into Cura's datatypes.xml).  As far as we can tell, there's no support for DEFAULT values in columns.  And while it does support foreign keys, it does so by declaring them in the column definition.  We didn't do that part yet, since it seemed to required more changes to normal ANSI than we wanted to make at this time.