[Version 2.0.4 Build 5] Corrected DAL problem reported by Rainer (COMSYS).
In the method AbstractTable.getRecordByEquality(), a local reference "stmt" of type PreparedStatement was initialized to null (line 454). This reference is passed to the method AbstractTable.getResultSet () (line 457). Inside this method, a PreparedStatement object is created and the passed-in is set to point to this object. After AbstractTable.getResultSet() returns to AbstractTable.getRecordByEquality (), an attempt is made to call stmt.close() (line 476). Unfortunately, stmt.close() is never called, because in
this method it still points to null. Corrected the problem by using rs.getStatement() and closing that statement instead.