Added "default" to the array of allowed column macro values.
Because the DAL generates the default value for inserts and updates when a value is not present, in some cases the value might need to be database specific.  An instance of this is a date colum: in Oracle, the function is "sysdate", while in SQL Server it is "getdate()"  By paramterizing the default value, it is possible to create a truly database-independent tabletype that can then be implemented in a database-specific manner.  Here is an example using the Default tabletype, abstracted one level to allow for this functionality:

<tabletype name="_Default">
  <param name="sysdatefunc">sysdate</param>
	<param name="record-status-enum">Record_Status</param>
	<column name="cr_stamp" type="stamp" required="dbms" default="$sysdatefunc$" descr="Date/time record was created"/>
	<column name="cr_person_id" lookupref="Person.person_id" descr="The user that created the record"/>
	<column name="cr_org_id" lookupref="Org.org_id" descr="The org that created the record (the one the cr_user_id was logged-into)"/>
	<column name="record_status_id" lookupref="$record-status-enum$" default="0" required="yes" descr="The status of a row in the table"/>
</tabletype>

<tabletype name="Default" type="_Default">
  <param name="sysdatefunc">getdate()</param>
</tabletype>