[Release 2.1.6]
* Fixed problem with conditionals when apply-flag was set to hidden
* Added new "hide readonly fields when values are null" feature so that QuerySelectDialogs can show criterias without looking too ugly.

NOTE: need to thoroughly test all dialogs because a subtle but important change was made to a methods called DialogField.isInputHidden() and DialogField.isVisible().

For the "hide readonly fields when values are null" there are two options:

1) When a field is set to read-only (like in confirm or viewing query select dialog criteria) then the value can be checked to see if it's null (or zero length string) and the field will automatically be HIDDEN.

2) When a field is set to read-only (like in confirm or viewing query select dialog criteria) then the value can be checked to see if it's null (or zero length string) and the field will automatically be INVISIBLE.

These flags can be set in a PER DIALOG or PER FIELD mode. When setting it per dialog, it applies to all fields. When setting it per field, it applies only to specific fields.

The new <dialog> attrbutes are
* "read-only-fields-hidden-unless-have-data" which can be set to "yes"
* "read-only-fields-invisible-unless-have-data" which can be set to "yes"

The new <field> attributes are
* read-only-hidden-unless-has-data which can be set to "yes"
* read-only-invisible-unless-has-data which can be set to "yes"