DBMS-Specific SQL Defaults --------------------------
This is a major reworking of the way Sparx handles SQL.  New is the capability
to specify SQL specific to a particular DBMS e.g. Oracle or PostgreSQL etc.
In previous versions, this capability was there but required a multi step
approach that included switching a datasource, changing all SQL to correspond
to the new DBMS (e.g. when switching from oracle to postgres, you would need
to switch all sysdate's to now()'s), recompiling the application to rebuild the
DAL with the new SQL in place etc.

Now it is possible to specify dbms specific SQL in the schema directly.  This
creates a DAL that is aware of all the DBMS specific SQL specified in the schema.
This means that effort to switch databases comes down to the simple step of
switching datasources.  No need to modify the schema (after adding dbms-specific
attributes) or recompile the DAL.

The only new tag that is in place as of right now is the <default> tag.  The
previously used default attribute inside a column or data type is now deprecated.
As an example of how to specify a default value for a column, you can look at
the following code:

<column name="cr_stamp" type="stamp" required="dbms" descr="Date/time record was created">
	<default dbms="oracle">sysdate</default>
	<default dbms="postgres">now()</default>
</column>

<column name="rec_stat_id" lookupref="$record-status-enum$" required="yes" descr="The status of a row in the table">
	<default dbms="oracle">0</default>
	<default dbms="postgres">'0'</default>
</column>

As you can see, the default value for the cr_stamp column can be sysdate (for oracle)
or now() (for postgresql).  Sparx chooses the appropriate value automatically as long
as you have provided these defaults.

These default values can also be set at the datatype level (e.g. for the stamp datatype)
and these will then take effect for columns of type stamp.

Java Method: getDBMSName()
--------------------------

This method is added to all DatabasePolicy objects and it returns a String containing
the name of the database (that the policy is written for) in the form that the dbms
attribute of the default tag (see example above) expects it.

Attribute: allow-sql-expr
-------------------------

Another addition is the new attribute allow-sql-expr.  If its value is set to "no" for
any column or data type, it marks that column as one that cannot accept default values
and that must have bind parameters when inserting or updating records.


Child Tag: java-default
-----------------------

The java-default child tag allows developers to specify the default Java value that a
data type will have when it is generated in the DAL.