[Release 2.1.19] Initial implementation the field persistence mechanism. Now, any field that has the "persist='yes'" attribute set will automatically create and read a cookie that will alow the dialog to remember values. Early testing completed, further testing required.
example:

<field.text name="abc" persist="yes"/>

Setting the persist flag to yes will allow the browser the "remember" the value of the field and automatically bring it up the next time the dialog is displayed.

The NextActions field in the Director is set to automatically persist, allowing a kind of basic workflow memory to be created.