[Version 2.1.25] Added new capability for schema columns and datatypes to define their own dialog fields separately and then aggregate them in a <dialog> tag in normal dialogs or special table dialogs. Here are the capabilities:
1) define <field.xxx> within <datatype> and <column> tags like this:

<datatype name="text">
    <field.text size="20" max-length="20">
    ....

<column name="gender" lookupref="Gender">
    <field.select caption="Caption" .../>

Some intelligence is built-in to the <field.xxx> definitions. For example:
1) If no "name" is specified, the name of the column is used
2) If no "caption" is specified, a useful caption is generated
3) If no "hint" is specified, the column description becomes the hint
4) All other attributes like primarykey, required, etc are inherited from the <column> and <datatype>
5) Any attributes specified in the <field.table-column> tag are carried into the <field.xxx> pulled in from the <column> tag.
6) Since this is an experimental feature right now, I am looking for other nice rules that can be implemented.

2) Those <field.xxx> may then referenced directly inside <dialog> tags using the new <field.table-column> tag like this:

<dialog name="xxx">
    <field.table-column table="abc" column="def"/>
    or
    <field.table-column table="abc" column="*"/>
    of
    <field.table-column table="abc" column="/A.*/"/>

In the first case, a <field.xxx> will be pulled in directly from the schema column called 'def' in the table 'abc'. In the second case, all columns that have <field.xxx> definitions in table 'abc' will be pulled in. In the third case, a perl regular expression is used to grab all columns that start with 'A' and have a <field.xxx> defined within.


3) Tables may now define a dialog associated with the table and those tables can be used in the DialogManager like this:

<table name="xxx">
    <dialog>
        <field.table-column column="*"/>
    </dialog>
</table>

or

<table name="xxx">
    <dialog name="me" class="you">
        ... any "normal" field
    </dialog>
</table>

to use these new table dialogs, simply create a reference to the dialog like this in WEB-INF/ui/dialogs.xml:

<dialogs package="whatever">
    <table-dialog table="Table_Name"/>
    or
    <table-dialog table="*"/>

This way you can define a single package and pull in one or more table dialogs automatically. All the dialogs will be documented and called as usual but will be defined near their tables, columns, and datatypes -- a big time savings.

In future releases of Sparx, automatic data management (insert/update/delete) functionality will also be provided for schema tables.